+++ lib/core/settings.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: non_constant_identifier_names
 
 
+import 'package:accredit/domain/models/application_info.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:accredit/dal/local/local_source_adapter.dart';
 
@@ -20,7 +21,20 @@ class Settings {
 
   final String imgsPath = "lib/presentation/assets/img/";
 
+  final String applicationInfo = ApplicationInfo(
+    name: "Accredit",
+    description: "Accredit is a secure and user-friendly platform for getting certifications.",
+    logoImageUrl: "https://example.com/logo.png",
+    urlApp: "http://localhost:1165",
+    twoFaAuth: false,
+    primaryColor: "#3498db",
+    secondaryColor: "#2ecc71",
+    tertiaryColor: "#e74c3c",
+    quartaryColor: null,
+    createdAt: DateTime.now().toIso8601String(),
+  ).toJson().toString();
 
+  final String AUTH_PARAM_KEY_NAME = "tokenized";