+++ CHANGELOG.md
@@ -37,6 +37,7 @@ All notable changes to this project will be documented in this file.
 - *(pdf)* Added a counter and limiter
 - *(pdf)* Adjusted screen for ai injection error
 - *(ui)* Modified the main color to purple to be more similar
+- *(sync)* Created initial widgets for auth apps
 
 ### 🐛 Bug Fixes